src/api/lib/netutils/requirements.txt
nameko
requests
Injector
python-dateutil
psycopg2
asynqp
pyserial-asyncio
